# perf: TopN Operator and Optimization