--- a/src/daft-local-execution/src/pipeline.rs
+++ b/src/daft-local-execution/src/pipeline.rs
@@ -14,7 +14,7 @@ use daft_dsl::{join::get_common_join_cols, resolved_col};
 use daft_local_plan::{
     ActorPoolProject, Concat, CrossJoin, EmptyScan, Explode, Filter, HashAggregate, HashJoin,
     InMemoryScan, Limit, LocalPhysicalPlan, MonotonicallyIncreasingId, PhysicalWrite, Pivot,
-    Project, Sample, Sort, UnGroupedAggregate, Unpivot, WindowPartitionAndDynamicFrame,
+    Project, Sample, Sort, TopN, UnGroupedAggregate, Unpivot, WindowPartitionAndDynamicFrame,
     WindowPartitionAndOrderBy, WindowPartitionOnly,
 };
 use daft_logical_plan::{stats::StatsState, JoinType};
@@ -49,6 +49,7 @@ use crate::{
         pivot::PivotSink,
         sort::SortSink,
         streaming_sink::StreamingSinkNode,
+        top_n::TopNSink,
         window_partition_and_dynamic_frame::WindowPartitionAndDynamicFrameSink,
         window_partition_and_order_by::WindowPartitionAndOrderBySink,
         window_partition_only::WindowPartitionOnlySink,
@@ -391,6 +392,24 @@ pub fn physical_plan_to_pipeline(
             let sort_sink = SortSink::new(sort_by.clone(), descending.clone(), nulls_first.clone());
             let child_node = physical_plan_to_pipeline(input, psets, cfg)?;
             BlockingSinkNode::new(Arc::new(sort_sink), child_node, stats_state.clone()).boxed()
+        }
+        LocalPhysicalPlan::TopN(TopN {
+            input,
+            sort_by,
+            descending,
+            nulls_first,
+            limit,
+            stats_state,
+            ..
+        }) => {
+            let sink = TopNSink::new(
+                sort_by.clone(),
+                descending.clone(),
+                nulls_first.clone(),
+                *limit as usize,
+            );
+            let child_node = physical_plan_to_pipeline(input, psets, cfg)?;
+            BlockingSinkNode::new(Arc::new(sink), child_node, stats_state.clone()).boxed()
         }
         LocalPhysicalPlan::MonotonicallyIncreasingId(MonotonicallyIncreasingId {
             input,