--- a/src/daft-local-execution/src/pipeline.rs
+++ b/src/daft-local-execution/src/pipeline.rs
@@ -403,6 +403,7 @@ pub fn physical_plan_to_pipeline(
             ..
         }) => {
             let sink = TopNSink::new(
+                input.schema(),
                 sort_by.clone(),
                 descending.clone(),
                 nulls_first.clone(),