--- a/src/daft-local-execution/src/sinks/mod.rs
+++ b/src/daft-local-execution/src/sinks/mod.rs
@@ -11,6 +11,7 @@ pub mod outer_hash_join_probe;
 pub mod pivot;
 pub mod sort;
 pub mod streaming_sink;
+pub mod top_n;
 pub mod window_base;
 pub mod window_partition_and_dynamic_frame;
 pub mod window_partition_and_order_by;