--- a/src/daft-local-execution/src/sinks/top_n.rs
+++ b/src/daft-local-execution/src/sinks/top_n.rs
@@ -1,6 +1,7 @@
 use std::sync::Arc;
 
 use common_error::DaftResult;
+use daft_core::prelude::SchemaRef;
 use daft_dsl::ExprRef;
 use daft_micropartition::MicroPartition;
 use itertools::Itertools;
@@ -12,57 +13,61 @@ use super::blocking_sink::{
 };
 use crate::ExecutionTaskSpawner;
 
+/// Parameters for the TopN that both the state and sinker need
 struct TopNParams {
+    // Sort By Parameters
     sort_by: Vec<ExprRef>,
     descending: Vec<bool>,
     nulls_first: Vec<bool>,
+    // Limit Parameters
     limit: usize,
 }
 
-enum TopNStatus {
-    Building,
+/// Current status of the TopN operation
+enum TopNState {
+    /// Operator is still collecting input and truncating to limit
+    Building(Arc<MicroPartition>),
+    /// Operator has finished collecting all input and ready to produce output
     Done,
 }
 
-struct TopNState {
-    top_values: Option<Arc<MicroPartition>>,
-    status: TopNStatus,
-}
-
 impl TopNState {
+    /// Process a new micro-partition and update the top N values
     fn push(&mut self, part: Arc<MicroPartition>, params: &TopNParams) {
-        if matches!(self.status, TopNStatus::Building) {
-            let top_rows_partition = part
-                .sort(&params.sort_by, &params.descending, &params.nulls_first)
-                .unwrap()
-                .slice(0, params.limit)
-                .unwrap();
-
-            let concat = if self.top_values.is_some() {
-                MicroPartition::concat([self.top_values.as_ref().unwrap(), &top_rows_partition])
-                    .unwrap()
-            } else {
-                top_rows_partition
-            };
-            let sorted = Arc::new(
-                concat
-                    .sort(&params.sort_by, &params.descending, &params.nulls_first)
-                    .unwrap(),
-            );
-            self.top_values = Some(sorted.slice(0, params.limit).unwrap().into());
-        } else {
-            panic!("TopNSink should be in Building state");
-        }
-    }
-
-    fn finalize(&mut self) -> Arc<MicroPartition> {
-        let res = if matches!(self.status, TopNStatus::Building) {
-            self.top_values.as_ref().unwrap()
-        } else {
+        let Self::Building(ref mut top_values) = self else {
             panic!("TopNSink should be in Building state");
         };
-        self.status = TopNStatus::Done;
-        res.clone()
+
+        // First find the top N values in the input micro-partition
+        let top_input_rows = part
+            .sort(&params.sort_by, &params.descending, &params.nulls_first)
+            .unwrap() // TODO: Where to catch errors
+            .slice(0, params.limit)
+            .unwrap();
+
+        // Then combine with existing top N values to have 2*N values
+        let concated = MicroPartition::concat([top_values, &top_input_rows]).unwrap();
+
+        // Re-sort the combined partition to get the top N values
+        // TODO: Use merge-sort merge algorithm to combine the two sorted partitions
+        let sorted = Arc::new(
+            concated
+                .sort(&params.sort_by, &params.descending, &params.nulls_first)
+                .unwrap(),
+        );
+
+        *top_values = sorted.slice(0, params.limit).unwrap().into();
+    }
+
+    /// Finalize the TopN operation and return the top N values
+    fn finalize(&mut self) -> Arc<MicroPartition> {
+        let Self::Building(top_values) = self else {
+            panic!("TopNSink should be in Building state");
+        };
+
+        let top_values = top_values.clone();
+        *self = Self::Done;
+        top_values
     }
 }
 
@@ -73,17 +78,20 @@ impl BlockingSinkState for TopNState {
 }
 
 pub struct TopNSink {
+    input_schema: SchemaRef,
     params: Arc<TopNParams>,
 }
 
 impl TopNSink {
     pub fn new(
+        input_schema: &SchemaRef,
         sort_by: Vec<ExprRef>,
         descending: Vec<bool>,
         nulls_first: Vec<bool>,
         limit: usize,
     ) -> Self {
         Self {
+            input_schema: input_schema.clone(),
             params: Arc::new(TopNParams {
                 sort_by,
                 descending,
@@ -179,9 +187,8 @@ impl BlockingSink for TopNSink {
     }
 
     fn make_state(&self) -> DaftResult<Box<dyn BlockingSinkState>> {
-        Ok(Box::new(TopNState {
-            top_values: None,
-            status: TopNStatus::Building,
-        }))
+        Ok(Box::new(TopNState::Building(Arc::new(
+            MicroPartition::empty(Some(self.input_schema.clone())),
+        ))))
     }
 }