--- a/src/daft-local-execution/src/sinks/top_n.rs
+++ b/src/daft-local-execution/src/sinks/top_n.rs
@@ -1,7 +1,6 @@
 use std::sync::Arc;
 
 use common_error::DaftResult;
-use daft_core::prelude::SchemaRef;
 use daft_dsl::ExprRef;
 use daft_micropartition::MicroPartition;
 use itertools::Itertools;
@@ -25,47 +24,30 @@ struct TopNParams {
 
 /// Current status of the TopN operation
 enum TopNState {
-    /// Operator is still collecting input and truncating to limit
-    Building(Arc<MicroPartition>),
+    /// Operator is still collecting input
+    Building(Vec<Arc<MicroPartition>>),
     /// Operator has finished collecting all input and ready to produce output
+    /// by doing a final sort + limit
     Done,
 }
 
 impl TopNState {
     /// Process a new micro-partition and update the top N values
-    fn push(&mut self, part: Arc<MicroPartition>, params: &TopNParams) {
+    fn append(&mut self, part: Arc<MicroPartition>) {
         let Self::Building(ref mut top_values) = self else {
             panic!("TopNSink should be in Building state");
         };
 
-        // First find the top N values in the input micro-partition
-        let top_input_rows = part
-            .sort(&params.sort_by, &params.descending, &params.nulls_first)
-            .unwrap() // TODO: Where to catch errors
-            .slice(0, params.limit)
-            .unwrap();
-
-        // Then combine with existing top N values to have 2*N values
-        let concated = MicroPartition::concat([top_values, &top_input_rows]).unwrap();
-
-        // Re-sort the combined partition to get the top N values
-        // TODO: Use merge-sort merge algorithm to combine the two sorted partitions
-        let sorted = Arc::new(
-            concated
-                .sort(&params.sort_by, &params.descending, &params.nulls_first)
-                .unwrap(),
-        );
-
-        *top_values = sorted.slice(0, params.limit).unwrap().into();
+        top_values.push(part);
     }
 
     /// Finalize the TopN operation and return the top N values
-    fn finalize(&mut self) -> Arc<MicroPartition> {
+    fn finalize(&mut self) -> Vec<Arc<MicroPartition>> {
         let Self::Building(top_values) = self else {
             panic!("TopNSink should be in Building state");
         };
 
-        let top_values = top_values.clone();
+        let top_values = std::mem::take(top_values);
         *self = Self::Done;
         top_values
     }
@@ -78,20 +60,17 @@ impl BlockingSinkState for TopNState {
 }
 
 pub struct TopNSink {
-    input_schema: SchemaRef,
     params: Arc<TopNParams>,
 }
 
 impl TopNSink {
     pub fn new(
-        input_schema: &SchemaRef,
         sort_by: Vec<ExprRef>,
         descending: Vec<bool>,
         nulls_first: Vec<bool>,
         limit: usize,
     ) -> Self {
         Self {
-            input_schema: input_schema.clone(),
             params: Arc::new(TopNParams {
                 sort_by,
                 descending,
@@ -119,7 +98,17 @@ impl BlockingSink for TopNSink {
                         .as_any_mut()
                         .downcast_mut::<TopNState>()
                         .expect("TopNSink should have top_n state");
-                    top_n_state.push(input, &params);
+
+                    // Find the top N values in the input micro-partition
+                    let top_input_rows = input.top_n(
+                        &params.sort_by,
+                        &params.descending,
+                        &params.nulls_first,
+                        params.limit,
+                    )?;
+
+                    // Append to the collection of existing top N values
+                    top_n_state.append(Arc::new(top_input_rows));
                     Ok(BlockingSinkStatus::NeedMoreInput(state))
                 },
                 Span::current(),
@@ -137,7 +126,7 @@ impl BlockingSink for TopNSink {
         spawner
             .spawn(
                 async move {
-                    let parts = states.into_iter().map(|mut state| {
+                    let parts = states.into_iter().flat_map(|mut state| {
                         let state = state
                             .as_any_mut()
                             .downcast_mut::<TopNState>()
@@ -187,8 +176,6 @@ impl BlockingSink for TopNSink {
     }
 
     fn make_state(&self) -> DaftResult<Box<dyn BlockingSinkState>> {
-        Ok(Box::new(TopNState::Building(Arc::new(
-            MicroPartition::empty(Some(self.input_schema.clone())),
-        ))))
+        Ok(Box::new(TopNState::Building(vec![])))
     }
 }