--- a/src/daft-micropartition/src/ops/mod.rs
+++ b/src/daft-micropartition/src/ops/mod.rs
@@ -9,5 +9,4 @@ mod pivot;
 mod slice;
 mod sort;
 mod take;
-mod top_n;
 mod unpivot;