## Changes Made

This PR adds a new TopN operator that gets the N largest / smallest values from input based on some sorting properties. It is essentially a Sort + Limit but can be implemented more efficiently as 1 operator rather than 2. This PR includes:

* An extension to limit pushdown to detect Sort + Limit -> TopN
* A naive distributed implementation of TopN that just does Sort + Limit
* A more efficiently local implementation of TopN 

Performance results of running ClickBench Queries 24, 25, & 26 (6 runs each, skipping first) on a M4 Max MacBook Pro w/ 14 cores & 32 GB memory:

| Query |	Before | After | Improvement |
| ------ | ------- | ----- | ------------- |
| Q24 | 1.43s | 0.746s | 1.91x |
| Q25 | 1.55s | 0.745s | 2.08x |
| Q26 | 1.54s | 0.732s | 2.10x |

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to `docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130 for docs review)