@colin-ho 2025-05-08
> @colin-ho In Python or Rust? I found existing top-n tests in Python

Ah ok, yeah I found them in cookbook. Those are quite old though, and we've added new options on sort like 'nulls_first' since then. Could you add a few top-n tests in tests/dataframe/test_sort.py, including some parametrized by `nulls_first`?