@srilman 2025-05-08
> Ah ok, yeah I found them in cookbook. Those are quite old though, and we've added new options on sort like 'nulls_first' since then. Could you add a few top-n tests in tests/dataframe/test_sort.py, including some parametrized by nulls_first?

I think the reason why there are no nulls_first tests is because Pandas doesn't have the option, at least for the granularity that Daft does. Any other tests you think I should add?