@colin-ho 2025-05-10
> Also whats the difference between the cookbook and dataframe tests?

cookbook tests are for 'real life' workloads